requirements.api.txt
Django==6.0.1
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.7.0
django-cors-headers==4.4.0